# openapi.yaml
openapi: 3.0.3
info:
  title: Quản lý CLB — API
  description: >
    API của Telegram Mini App quản lý CLB (điểm danh, chi phí, sao kê).
    Tham khảo thiết kế tại thiet-ke-mini-app-clb.md.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://your-app.vercel.app
    description: Production (thay bằng NEXT_PUBLIC_APP_URL thật)
...
security:
  - cookieAuth: []
paths:
  /api/auth/telegram:
    post:
      tags: [auth]
      summary: Đăng nhập admin bằng Telegram WebApp initData
      description: >
        Verify initData bằng HMAC-SHA256 (bot token), tìm member tương ứng — chỉ
        role=admin và status=active mới được cấp session. Set cookie clb_session (JWT, httpOnly).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData:
                  type: string
                  description: window.Telegram.WebApp.initData lấy từ FE
      responses:
        "200":
          description: Đăng nhập thành công, cookie session đã được set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  member:
                    type: object
                    properties:
                      id:
                        type: string
                      full_name:
                        type: string
                      role:
                        type: string
        "400":
          description: Thiếu initData
        "401":
          description: initData không hợp lệ (sai chữ ký hoặc hết hạn)
        "403":
          description: telegram_id không thuộc admin đang active
        "500":
          description: Server chưa cấu hình TELEGRAM_BOT_TOKEN
  /api/auth/join:
    post:
      tags: [auth]
      summary: Tham gia với tư cách thành viên / gửi yêu cầu làm admin
      description: >
        Dùng chung cho 2 nút ở màn hình đăng nhập khi bị chặn ("Tham gia với tư
        cách thành viên" và "Gửi yêu cầu làm admin"). Verify initData bằng
        HMAC-SHA256 (bot token). Nếu telegram_id chưa có trong bảng member thì
        tạo mới với role=member, status=active. Không tự gán role=admin — quản
        trị viên hệ thống phải vào DB nâng role thủ công cho tài khoản này thì
        mới đăng nhập được vào /api/auth/telegram.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData:
                  type: string
                  description: window.Telegram.WebApp.initData lấy từ FE
      responses:
        "200":
          description: Đã tạo member (hoặc đã tồn tại từ trước)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  alreadyExists:
                    type: boolean
        "400":
          description: Thiếu initData
        "401":
          description: initData không hợp lệ (sai chữ ký hoặc hết hạn)
        "500":
          description: Server chưa cấu hình TELEGRAM_BOT_TOKEN
...
  /api/members/{id}:
    get:
      tags: [members]
      summary: Chi tiết 1 thành viên — thông tin cơ bản, số buổi tập theo tháng lọc, sao kê từng tháng
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Tháng cần xem số buổi tham gia (1-12). Mặc định tháng hiện tại (giờ VN).
        - name: year
          in: query
          required: false
          schema:
            type: integer
          description: Năm cần xem, đi kèm `month`. Mặc định năm hiện tại (giờ VN).
      responses:
        "200":
          description: Chi tiết thành viên
          content:
            application/json:
              schema:
                type: object
                properties:
                  member:
                    $ref: "#/components/schemas/Member"
                  monthKey:
                    type: string
                    description: "Tháng đang lọc, dạng YYYY-MM"
                  sessionsAttended:
                    type: integer
                    description: Số buổi tham gia (answer=present) trong monthKey
                  statements:
                    type: array
                    description: Toàn bộ sao kê của thành viên, mọi tháng, mới nhất trước
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        month:
                          type: string
                        total_sessions:
                          type: integer
                        total_amount:
                          type: integer
                        status:
                          type: string
                          enum: [pending, paid_reported, approved]
        "401":
          description: Chưa đăng nhập
        "404":
          description: Không tìm thấy thành viên
    patch:
      tags: [members]
      summary: Sửa thông tin thành viên (hiện chỉ hỗ trợ group chat riêng)
      description: >
        Set/xoá `statement_chat_id` — group chat riêng của thành viên (admin + member này)
        dùng để tự động gửi noti sao kê hàng tháng. Truyền `null` để xoá.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                statement_chat_id:
                  type: integer
                  nullable: true
      responses:
        "200":
          description: Member đã cập nhật
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Payload không hợp lệ
        "401":
          description: Chưa đăng nhập
        "404":
          description: Không tìm thấy thành viên
    delete:
      tags: [members]
      summary: Hủy thành viên (soft delete — set del_flag=true, status=inactive)
      description: >
        Không xoá dữ liệu, chỉ ẩn khỏi danh sách và thu hồi quyền truy cập (kể cả
        nếu là admin). Không cho phép admin tự hủy chính mình.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Đã hủy thành viên
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: Không thể tự hủy chính mình
        "401":
          description: Chưa đăng nhập
        "404":
          description: Không tìm thấy thành viên
...
  /api/cron/attendance/{entryId}:
    post:
      tags: [cron]
      summary: Job điểm danh của 1 lịch tập — được job cron-job.org riêng của mục đó gọi
      description: >
        Mỗi mục weekly_schedule có 1 job cron-job.org riêng (tự tạo/cập nhật/xoá khi lưu
        /api/settings, xem cronjob_id), kích hoạt đúng mốc "trước giờ tập N tiếng"
        (reminder_hours_before). Khi gọi: tạo Session của lần diễn ra kế tiếp nếu chưa có, rồi gửi
        poll điểm danh vào main_group_chat_id. Idempotent qua poll_sent_at.
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: string
          description: _id của mục trong settings.weekly_schedule
      security:
        - cronBearer: []
      responses:
        "200":
          description: Đã đảm bảo Session tồn tại và đã gửi (hoặc đã từng gửi) poll
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  sessionId:
                    type: string
                  pollSent:
                    type: boolean
        "401":
          description: Sai hoặc thiếu CRON_SECRET
        "404":
          description: Không tìm thấy entryId trong weekly_schedule (job cũ, đã bị xoá)
  /api/settings:
    get:
      tags: [settings]
      summary: Lấy cấu hình CLB hiện tại (tự tạo document mặc định nếu chưa có)
      responses:
        "200":
          description: Cấu hình hiện tại
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Chưa đăng nhập
    put:
      tags: [settings]
      summary: Cập nhật cấu hình CLB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Cấu hình sau khi cập nhật
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Chưa đăng nhập
...
  /api/statements/{id}:
    patch:
      tags: [statements]
      summary: Admin chỉnh tay 1 sao kê (VD tính sai chi phí)
      description: Sửa total_sessions, total_amount và/hoặc status. Dùng từ trang chi tiết thành viên.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                total_sessions:
                  type: integer
                total_amount:
                  type: integer
                status:
                  type: string
                  enum: [pending, paid_reported, approved]
      responses:
        "200":
          description: Sao kê sau khi sửa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyStatement"
        "400":
          description: Payload không hợp lệ
        "401":
          description: Chưa đăng nhập
        "404":
          description: Không tìm thấy sao kê
    delete:
      tags: [statements]
      summary: Xoá 1 sao kê
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Đã xoá
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          description: Chưa đăng nhập
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: clb_session
      description: JWT session cookie (httpOnly), cấp sau khi đăng nhập qua /api/auth/telegram.
    telegramSecretToken:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Do Telegram tự gửi lên khi gọi webhook, khớp với TELEGRAM_WEBHOOK_SECRET.
    cronBearer:
      type: http
      scheme: bearer
      description: "Header Authorization: Bearer <CRON_SECRET>, do cron-job.org gửi lên."
  schemas:
    Settings:
      type: object
      properties:
        club_name:
          type: string
        main_group_chat_id:
          type: integer
          description: chat_id nhóm ① (chính)
        admin_group_chat_id:
          type: integer
          description: chat_id nhóm ② (Ban quản trị)
        weekly_schedule:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                description: Định danh entry — dùng làm {entryId} của /api/cron/attendance/{entryId}
              weekday:
                type: integer
                minimum: 0
                maximum: 6
                description: 0 = Chủ nhật ... 6 = Thứ 7
              start_time:
                type: string
                example: "18:30"
              end_time:
                type: string
                example: "20:00"
              cronjob_id:
                type: integer
                description: id job tương ứng trên cron-job.org (tự quản lý khi lưu settings)
        required_participants:
          type: integer
          description: >
            Số người tham gia kỳ vọng mỗi buổi — làm min_required mặc định cho buổi tự tạo và mốc
            so sánh "đủ/thiếu người"
        fixed_cost_per_session:
          type: integer
          description: Chi phí cố định phát sinh mỗi buổi (thuê sân, ...), cộng vào chi phí vật phẩm khi tính sao kê
        reminder_hours_before:
          type: integer
          description: Gửi poll trước giờ tập bao nhiêu tiếng
        cost_survey_minutes_after:
          type: integer
          description: Nhắc nhập chi phí sau khi kết thúc bao nhiêu phút
        monthly_settlement_day:
          type: integer
          minimum: 1
          maximum: 28
        last_monthly_settlement_run:
          type: string
          description: "Tháng đã chốt sao kê gần nhất, dạng YYYY-MM (dùng chống chạy trùng)"
...
    MonthlyStatement:
      type: object
      properties:
        _id:
          type: string
        member_id:
          type: string
        month:
          type: string
          example: "2026-07"
        total_sessions:
          type: integer
        total_amount:
          type: number
        status:
          type: string
          enum: [pending, paid_reported, approved]
        paid_reported_at:
          type: string
          format: date-time
        approved_at:
          type: string
          format: date-time
        approved_by:
          type: string
    Member:
      type: object
      properties:
        _id:
          type: string
        telegram_id:
          type: integer
        full_name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, member]
        status:
          type: string
          enum: [active, inactive]
        del_flag:
          type: boolean
          description: true nếu admin đã hủy thành viên này (soft delete)
        statement_chat_id:
          type: integer
          description: chat_id nhóm ③ riêng của member này
        joined_at:
          type: string
          format: date-time